done backend take valua

+++ hardware/src/flame_setup.cpp
@@ -11,8 +11,8 @@ void vTaskFlame(void *pvParameter) {
             if (isnan(rawValue1) || isnan(rawValue2)) {
                 Serial.print("\nValue is unreadable!"); 
             } else {
-                flame[0] = (1.0 - rawValue1/4095.0) * 100;
-                flame[1] = (1.0 - rawValue2/4095.0) * 100;
+                flame[0] = (1.0 - (float)rawValue1/4095.0) * 100;
+                flame[1] = (1.0 - (float)rawValue2/4095.0) * 100;
 
                 Serial.printf("\nRaw Value: %.2f | Flame Detect: %.2f", rawValue1, flame[0]); 
                 Serial.printf("\nRaw Value: %.2f | Flame Detect: %.2f", rawValue2, flame[1]);
